+++ src/architecture/msgPayloadDefC/NavAttMsgPayload.h
@@ -22,10 +22,10 @@
 
 /*! @brief Structure used to define the output definition for attitude guidance*/
 typedef struct {
-    double timeTag;          //!< [s]   Current vehicle time-tag associated with measurements*/
-    double sigma_BN[3];      //!<       Current spacecraft attitude (MRPs) of body relative to inertial */
-    double omega_BN_B[3];    //!< [r/s] Current spacecraft angular velocity vector of body frame B relative to inertial frame N, in B frame components
-    double vehSunPntBdy[3];  //!<       Current sun pointing vector in body frame
+    float timeTag;          //!< [s]   Current vehicle time-tag associated with measurements*/
+    float sigma_BN[3];      //!<       Current spacecraft attitude (MRPs) of body relative to inertial */
+    float omega_BN_B[3];    //!< [r/s] Current spacecraft angular velocity vector of body frame B relative to inertial frame N, in B frame components
+    float vehSunPntBdy[3];  //!<       Current sun pointing vector in body frame
 }NavAttMsgPayload;
