+++ src/architecture/msgPayloadDefC/RWArrayConfigMsgPayload.h
@@ -25,10 +25,10 @@
 
 /*! @brief RW array configuration FSW msg */
 typedef struct{
-    double GsMatrix_B[3*MAX_EFF_CNT];   //!< [-]    The RW spin axis matrix in body frame components 
-    double JsList[MAX_EFF_CNT];         //!< [kgm2] The spin axis inertia for RWs
+    float GsMatrix_B[3*MAX_EFF_CNT];   //!< [-]    The RW spin axis matrix in body frame components 
+    float JsList[MAX_EFF_CNT];         //!< [kgm2] The spin axis inertia for RWs
     int    numRW;                       //!< [-]    The number of reaction wheels available on vehicle
-    double uMax[MAX_EFF_CNT];           //!< [Nm]   The maximum RW motor torque
+    float uMax[MAX_EFF_CNT];           //!< [Nm]   The maximum RW motor torque
 }RWArrayConfigMsgPayload;
