+++ src/architecture/msgPayloadDefC/SunlineFilterMsgPayload.h
@@ -30,11 +30,11 @@
 /*! @brief structure for filter-states output for the unscented kalman filter
  implementation of the sunline state estimator*/
 typedef struct {
-    double timeTag;                             //!< [s] Current time of validity for output 
-    double covar[SKF_N_STATES*SKF_N_STATES];    //!< [-] Current covariance of the filter
-    double state[SKF_N_STATES];                 //!< [-] Current estimated state of the filter
-    double stateError[SKF_N_STATES];            //!< [-] Current deviation of the state from the reference state
-    double postFitRes[MAX_N_CSS_MEAS];          //!< [-] PostFit Residuals
+    float timeTag;                             //!< [s] Current time of validity for output 
+    float covar[SKF_N_STATES*SKF_N_STATES];    //!< [-] Current covariance of the filter
+    float state[SKF_N_STATES];                 //!< [-] Current estimated state of the filter
+    float stateError[SKF_N_STATES];            //!< [-] Current deviation of the state from the reference state
+    float postFitRes[MAX_N_CSS_MEAS];          //!< [-] PostFit Residuals
     int numObs;                                 //!< [-] Valid observation count for this frame
 }SunlineFilterMsgPayload;
