+++ src/architecture/msgPayloadDefC/VehicleConfigMsgPayload.h
@@ -25,9 +25,9 @@
 
 /*! @brief Structure used to define a common structure for top level vehicle information*/
 typedef struct {
-    double ISCPntB_B[9];          //!< [kg m^2] Spacecraft Inertia 
-    double CoM_B[3];              //!< [m] Center of mass of spacecraft in body
-    double massSC;                //!< [kg] Spacecraft mass
+    float ISCPntB_B[9];          //!< [kg m^2] Spacecraft Inertia 
+    float CoM_B[3];              //!< [m] Center of mass of spacecraft in body
+    float massSC;                //!< [kg] Spacecraft mass
     uint32_t CurrentADCSState;    //!< [-] Current ADCS state for subsystem
 }VehicleConfigMsgPayload;
